+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarController.java
@@ -98,6 +98,12 @@ public class TablesSidebarController implements Initializable {
   private Button tablesBtn;
 
   @FXML
+  private Button povBtn;
+
+  @FXML
+  private Button pupBackBtn;
+
+  @FXML
   private HBox popperTitleButtonArea;
 
   @FXML
@@ -174,6 +180,16 @@ public class TablesSidebarController implements Initializable {
     try {
       SystemSummary systemSummary = Studio.client.getSystemService().getSystemSummary();
       new ProcessBuilder("explorer.exe", new File(systemSummary.getVisualPinballDirectory(), "Tables").getAbsolutePath()).start();
+    } catch (Exception e) {
+      LOG.error("Failed to open Explorer: " + e.getMessage(), e);
+    }
+  }
+
+  @FXML
+  private void onPupPack() {
+    try {
+      SystemSummary systemSummary = Studio.client.getSystemService().getSystemSummary();
+      new ProcessBuilder("explorer.exe", new File(systemSummary.getPinupSystemDirectory(), "PUPVideos").getAbsolutePath()).start();
     } catch (Exception e) {
       LOG.error("Failed to open Explorer: " + e.getMessage(), e);
     }
@@ -279,6 +295,8 @@ public class TablesSidebarController implements Initializable {
     scriptBtn.setVisible(client.getSystemService().isLocal());
     nvramExplorerBtn.setVisible(client.getSystemService().isLocal());
     tablesBtn.setVisible(client.getSystemService().isLocal());
+    povBtn.setVisible(client.getSystemService().isLocal());
+    pupBackBtn.setVisible(client.getSystemService().isLocal());
 
     try {
       FXMLLoader loader = new FXMLLoader(TablesSidebarAltSoundController.class.getResource("scene-tables-sidebar-altsound.fxml"));