+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/TablePopperMediaDialogController.java
@@ -795,6 +795,7 @@ public class TablePopperMediaDialogController implements Initializable, DialogCo
 
 
   private void refreshTableMediaView() {
+    this.helpBtn.setDisable(!PopperScreen.Loading.equals(screen));
     if (screen.equals(PopperScreen.Wheel)) {
       client.getImageCache().clearWheelCache();
     }